pyproject.toml
[tool.poetry]
name = "screenshot-service"
version = "0.1.0"
description = "Screenshot retrieval service for Giggityflix media streaming platform"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.27.1"
redis = "^5.0.1"
minio = "^7.2.3"
python-multipart = "0.0.19"
pydantic = "^2.7.1"
httpx = "^0.27.0"
confluent-kafka = "^2.3.0"
python-jose = "^3.3.0"
starlette = "^0.40.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.3.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"
freezegun = "^1.4.0"
moto = "^4.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: marks unit tests",
    "integration: marks integration tests",
    "functional: marks functional tests",
    "slow: marks tests as slow",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
#addopts = "--cov=src --cov-report=term-missing:skip-covered --cov-report=xml --cov-branch"